提交修改

+++ TCCarShare/Entity/Response/GeocoderResponse.cs
@@ -20,12 +20,12 @@ namespace TCCarShare.Entity.Response
         /// <summary>
         /// 纬度
         /// </summary>
-        public decimal Lat { get; set; }
+        public double Lat { get; set; }
 
         /// <summary>
         /// 经度
         /// </summary>
-        public decimal Lng { get; set; }
+        public double Lng { get; set; }
     }
 
     /// <summary>
